+++ hubdle-gradle-plugin/testFunctional/resources/config/format/apply-1/main/kotlin/Format.kt
@@ -2,5 +2,8 @@ package com.kotlin.jvm.sandbox.project
 
 class Format {
 
-    val hello = 1
+    val hello =
+
+
+        1
 }